requirements.txt
zeroconf
Pillow
requests